fix: critical bugs + code quality refactors
- Fix exports types condition for TS consumers
- Fix http.ts double-connect on same Client instance
- Fix watch mode error swallowing (silent poll failure)
- Fix --no-color detection broken by Commander argv mutation
- Extract security check helper (37 lines removed)
- Refactor server.ts from 756 → 55 lines (15 tool handler files)
- Add Zod schemas to discovery.ts, remove 9 as-casts
- Construct explicit types in validate.ts instead of as-casting
