+++ src/validate.ts
@@ -1,7 +1,29 @@
-import type { DiffArtifact, RunArtifact, TargetConfig } from "./types.js";
+import type {
+  CheckResult,
+  CheckStatus,
+  DiffArtifact,
+  DiffEntry,
+  DiffSummary,
+  EnvironmentSnapshot,
+  EvidenceSummary,
+  Gate,
+  HealthScore,
+  PerformanceMetrics,
+  ResponseChangeEntry,
+  RunArtifact,
+  RunSummary,
+  SchemaDriftEntry,
+  TargetConfig,
+  TargetSnapshot,
+} from "./types.js";
+import { SCHEMA_VERSION } from "./types.js";
 
-function isObject(value: unknown): value is Record<string, unknown> {
+export function isObject(value: unknown): value is Record<string, unknown> {
   return typeof value === "object" && value !== null && !Array.isArray(value);
+}
+
+function isStringArray(value: unknown): value is string[] {
+  return Array.isArray(value) && value.every((v) => typeof v === "string");
 }
 
 function requireString(obj: Record<string, unknown>, field: string, label: string): string {
@@ -37,42 +59,61 @@ function requireNumber(obj: Record<string, unknown>, field: string, label: strin
   return value;
 }
 
-function requireStatus(value: unknown, label: string): void {
+function requireGate(value: unknown, label: string): Gate {
+  if (value !== "pass" && value !== "fail") {
+    throw new Error(`${label} has invalid gate '${String(value)}'.`);
+  }
+  return value;
+}
+
+function requireStatus(value: unknown, label: string): CheckStatus {
   const statuses = new Set(["pass", "fail", "partial", "unsupported", "flaky", "skipped"]);
   if (typeof value !== "string" || !statuses.has(value)) {
     throw new Error(`${label} has invalid status '${String(value)}'.`);
   }
+  return value as CheckStatus;
 }
 
-function requireCheckId(value: unknown, label: string): void {
+function requireCheckId(value: unknown, label: string): string {
   const ids = new Set(["tools", "prompts", "resources", "tools-invoke", "security", "security-lite", "attack-sim", "conformance", "schema-quality"]);
   if (typeof value !== "string" || !ids.has(value)) {
     throw new Error(`${label} has invalid check id '${String(value)}'.`);
   }
+  return value;
 }
 
-function validateRunSummary(value: unknown): void {
+function validateRunSummary(value: unknown): RunSummary {
   if (!isObject(value)) {
     throw new Error("Run artifact is missing required field 'summary'.");
   }
-  if (value["gate"] !== "pass" && value["gate"] !== "fail") {
-    throw new Error("Run artifact summary has invalid gate.");
-  }
-  for (const field of ["total", "pass", "fail", "partial", "unsupported", "flaky", "skipped"]) {
-    requireNumber(value, field, "Run artifact summary");
-  }
+  const gate = requireGate(value["gate"], "Run artifact summary");
+  return {
+    gate,
+    total: requireNumber(value, "total", "Run artifact summary"),
+    pass: requireNumber(value, "pass", "Run artifact summary"),
+    fail: requireNumber(value, "fail", "Run artifact summary"),
+    partial: requireNumber(value, "partial", "Run artifact summary"),
+    unsupported: requireNumber(value, "unsupported", "Run artifact summary"),
+    flaky: requireNumber(value, "flaky", "Run artifact summary"),
+    skipped: requireNumber(value, "skipped", "Run artifact summary"),
+  };
 }
 
-function validateCheck(value: unknown, index: number): void {
+function validateCheck(value: unknown, index: number): CheckResult {
   if (!isObject(value)) {
     throw new Error(`Run artifact checks[${index}] must be an object.`);
   }
-  requireCheckId(value["id"], `Run artifact checks[${index}]`);
-  requireCheckId(value["capability"], `Run artifact checks[${index}] capability`);
-  requireStatus(value["status"], `Run artifact checks[${index}]`);
-  requireNumber(value, "durationMs", `Run artifact checks[${index}]`);
-  requireString(value, "message", `Run artifact checks[${index}]`);
-  requireArray(value, "evidence", `Run artifact checks[${index}]`);
+  const id = requireCheckId(value["id"], `Run artifact checks[${index}]`);
+  const capability = requireCheckId(value["capability"], `Run artifact checks[${index}] capability`);
+  const status = requireStatus(value["status"], `Run artifact checks[${index}]`);
+  return {
+    id: id as CheckResult["id"],
+    capability: capability as CheckResult["capability"],
+    status,
+    durationMs: requireNumber(value, "durationMs", `Run artifact checks[${index}]`),
+    message: requireString(value, "message", `Run artifact checks[${index}]`),
+    evidence: (Array.isArray(value["evidence"]) ? value["evidence"].filter(isObject) : []) as unknown as EvidenceSummary[],
+  };
 }
 
 function expandEnvValue(value: string, label: string): string {
@@ -174,30 +215,65 @@ export function validateRunArtifact(data: unknown): RunArtifact {
   if (data["artifactType"] !== "run") {
     throw new Error(`Expected a run artifact but got artifactType='${String(data["artifactType"])}'.`);
   }
-  requireString(data, "runId", "Run artifact");
-  requireString(data, "createdAt", "Run artifact");
-  requireString(data, "schemaVersion", "Run artifact");
-  requireString(data, "toolVersion", "Run artifact");
-  const checks = requireArray(data, "checks", "Run artifact");
+
+  const runId = requireString(data, "runId", "Run artifact");
+  const createdAt = requireString(data, "createdAt", "Run artifact");
+  const schemaVersion = requireString(data, "schemaVersion", "Run artifact");
+  const toolVersion = requireString(data, "toolVersion", "Run artifact");
+  const gate = requireGate(data["gate"], "Run artifact");
 
   if (!isObject(data["target"])) {
     throw new Error("Run artifact is missing required field 'target'.");
   }
-  requireString(data["target"], "targetId", "Run artifact target");
-  requireString(data["target"], "adapter", "Run artifact target");
+  const target: TargetSnapshot = {
+    targetId: requireString(data["target"], "targetId", "Run artifact target"),
+    adapter: requireString(data["target"], "adapter", "Run artifact target") as TargetSnapshot["adapter"],
+    command: typeof data["target"]["command"] === "string" ? data["target"]["command"] : "",
+    args: isStringArray(data["target"]["args"]) ? data["target"]["args"].filter((a): a is string => typeof a === "string") : [],
+    url: typeof data["target"]["url"] === "string" ? data["target"]["url"] : undefined,
+    cwd: typeof data["target"]["cwd"] === "string" ? data["target"]["cwd"] : undefined,
+    metadata: optionalStringRecord(data["target"]["metadata"], "Run artifact target metadata"),
+    serverVersion: typeof data["target"]["serverVersion"] === "string" ? data["target"]["serverVersion"] : undefined,
+    serverName: typeof data["target"]["serverName"] === "string" ? data["target"]["serverName"] : undefined,
+  };
+
   if (!isObject(data["environment"])) {
     throw new Error("Run artifact is missing required field 'environment'.");
   }
-  requireString(data["environment"], "platform", "Run artifact environment");
-  requireString(data["environment"], "nodeVersion", "Run artifact environment");
-  validateRunSummary(data["summary"]);
-  for (const [index, check] of checks.entries()) {
-    validateCheck(check, index);
-  }
+  const environment: EnvironmentSnapshot = {
+    platform: requireString(data["environment"], "platform", "Run artifact environment"),
+    nodeVersion: requireString(data["environment"], "nodeVersion", "Run artifact environment"),
+  };
 
-  // Structure validated above. The intermediate unknown cast is required because
-  // TypeScript can't narrow Record<string, unknown> to a specific interface.
-  return data as unknown as RunArtifact;
+  const summary = validateRunSummary(data["summary"]);
+  const checksRaw = requireArray(data, "checks", "Run artifact");
+  const checks: CheckResult[] = checksRaw.map((c, i) => validateCheck(c, i));
+
+  const healthScore: HealthScore | undefined = isObject(data["healthScore"])
+    ? data["healthScore"] as unknown as HealthScore
+    : undefined;
+
+  const performanceMetrics: PerformanceMetrics | undefined = isObject(data["performanceMetrics"])
+    ? data["performanceMetrics"] as unknown as PerformanceMetrics
+    : undefined;
+
+  const fatalError: string | undefined = typeof data["fatalError"] === "string" ? data["fatalError"] : undefined;
+
+  return {
+    artifactType: "run",
+    schemaVersion: schemaVersion as typeof SCHEMA_VERSION,
+    gate,
+    runId,
+    createdAt,
+    toolVersion,
+    target,
+    environment,
+    summary,
+    checks,
+    healthScore,
+    performanceMetrics,
+    fatalError,
+  };
 }
 
 export function validateDiffArtifact(data: unknown): DiffArtifact {
@@ -207,14 +283,50 @@ export function validateDiffArtifact(data: unknown): DiffArtifact {
   if (data["artifactType"] !== "diff") {
     throw new Error(`Expected a diff artifact but got artifactType='${String(data["artifactType"])}'.`);
   }
-  requireString(data, "baseRunId", "Diff artifact");
-  requireString(data, "headRunId", "Diff artifact");
-  requireString(data, "createdAt", "Diff artifact");
-  requireString(data, "schemaVersion", "Diff artifact");
 
-  if (!isObject(data["summary"])) {
+  const baseRunId = requireString(data, "baseRunId", "Diff artifact");
+  const headRunId = requireString(data, "headRunId", "Diff artifact");
+  const createdAt = requireString(data, "createdAt", "Diff artifact");
+  const schemaVersion = requireString(data, "schemaVersion", "Diff artifact");
+  const gate = requireGate(data["gate"], "Diff artifact");
+
+  const summaryObj = data["summary"];
+  if (!isObject(summaryObj)) {
     throw new Error("Diff artifact is missing required field 'summary'.");
   }
+  const summary: DiffSummary = {
+    regressions: requireNumber(summaryObj, "regressions", "Diff artifact summary"),
+    recoveries: requireNumber(summaryObj, "recoveries", "Diff artifact summary"),
+    unchanged: requireNumber(summaryObj, "unchanged", "Diff artifact summary"),
+    added: requireNumber(summaryObj, "added", "Diff artifact summary"),
+    removed: requireNumber(summaryObj, "removed", "Diff artifact summary"),
+    schemaDriftCount: optionalNumber(summaryObj, "schemaDriftCount", "Diff artifact summary"),
+    responseChangeCount: optionalNumber(summaryObj, "responseChangeCount", "Diff artifact summary"),
+    gate: requireGate(summaryObj["gate"], "Diff artifact summary"),
+  };
 
-  return data as unknown as DiffArtifact;
+  const regressions = Array.isArray(data["regressions"]) ? (data["regressions"] as DiffEntry[]) : [];
+  const recoveries = Array.isArray(data["recoveries"]) ? (data["recoveries"] as DiffEntry[]) : [];
+  const unchanged = Array.isArray(data["unchanged"]) ? (data["unchanged"] as DiffEntry[]) : [];
+  const added = Array.isArray(data["added"]) ? (data["added"] as DiffEntry[]) : [];
+  const removed = Array.isArray(data["removed"]) ? (data["removed"] as DiffEntry[]) : [];
+  const schemaDrift = Array.isArray(data["schemaDrift"]) ? (data["schemaDrift"] as SchemaDriftEntry[]) : undefined;
+  const responseChanges = Array.isArray(data["responseChanges"]) ? (data["responseChanges"] as ResponseChangeEntry[]) : undefined;
+
+  return {
+    artifactType: "diff",
+    schemaVersion: schemaVersion as typeof SCHEMA_VERSION,
+    gate,
+    baseRunId,
+    headRunId,
+    createdAt,
+    summary,
+    regressions,
+    recoveries,
+    unchanged,
+    added,
+    removed,
+    schemaDrift,
+    responseChanges,
+  };
 }
